Move yubikey-related commands to a separate file.
The goal is to centralize code for this feature, so we have fewer
[cfg(feature = "yubikey")] pragmas scattered around.
